refactor(utils): reorganize utility modules and imports
