+++ src/utils/constants.ts
@@ -1,8 +1,6 @@
 import { getModuleData } from '@inverter-network/abis'
 import { formatUnits } from 'viem'
 
-export const ADDRESS_ZERO = '0x0000000000000000000000000000000000000000'
-
 // Uint Max Supply
 export const UINT_MAX_SUPPLY =
   115792089237316195423570985008687907853269984665640564039457584007913129639935n
