+++ src/utils/index.ts
@@ -1,14 +1,14 @@
 import type { Entries } from 'type-fest-4'
 
-export * from './getJsType'
-export * from './handleError'
+export * from './get-js-type'
+export * from './handle-error'
 export * from './constants'
-export * from './processInputs'
-export * from './formatOutputs'
+export * from './process-inputs'
+export * from './format-outputs'
 export * from './external'
 
-export { default as tagProcessor } from './tagProcessor'
-export { default as handleOptions } from './handleOptions'
+export { default as tagProcessor } from './tag-processor'
+export { default as handleOptions } from './handle-options'
 
 export const getEntries = <T extends object>(obj: T): Entries<T> =>
   Object.entries(obj) as any
